Created Observable factory
There is now an object to lookup a city, fetch the results, and then emit an item of the result. It is published, so 2 things are handling results. One that simply logs, another that updates the DOM. Another stream is forking off of it, pushing a <p> to the DOM in the case of a city being hot.